braise v0.1.0 Braise.AdapterTemplate

Constructs a string representation of an Ember-CLI Adapter given the passed in JSON Schema Resource.

Functions

generate_from_resource(resource, template_string \\ "// DO NOT EDIT - this file was autogenerated by Braise\nimport DS from 'ember-data';\nimport Ember from 'ember';\n\nexport default DS.RESTAdapter.extend({\n host: \"<%= scheme %>://<%= host %>\",\n namespace: \"<%= path %>\",\n token: Ember.computed.alias('accessTokenWrapper.token'),\n <%= path_for_type %>\n headers: function() {\n return {\n 'AUTHORIZATION': 'Bearer ' + this.get('token')\n };\n }.property('token')\n});\n")

Returns a tuple that contains either a string representation of the adapter or an error message.

Examples

iex > resource = %Braise.Resource{definitions: %{"patients" =>{}},
                                  links: [%{"href" => "http://bizdev.biz/api/v1"}]}
iex > Braise.AdapterTemplate.generate_from_resource(resource)
{:ok,
 "import DS from 'ember-data';

import Ember from ‘ember’;

export default DS.RESTAdapter.extend({ host: “http://bizdev.biz”,

 namespace: "/api/v1",

token: Ember.computed.alias(‘accessTokenWrapper.token’),

 headers: function() {
return {
  'AUTHORIZATION': 'Bearer ' + this.get('token');

 };

}.property(‘token’) }); “

}

iex > Braise.AdapterTemplate.generate_from_resource(%{})
{:error, "Invalid JSON Schema"}

iex > Braise.AdapterTemplate.generate_from_resource(%Braise.Resource{definitions: %{"patients" => {}}, links: "Pirate Booty"})
{:error, "Invalid links portion of JSON Schema"}

iex > Braise.AdapterTemplate.generate_From_resource(%Braise.Resource{definitions: "Pirate Booty", links: [%{"href" => "http://piratebooty.biz/"}])
{:error, "Invalid definitions portion of JSON Schema"}

replace_path_for_type_variable(template, resource_name)

Replaces the path_for_type template variable if it is the passed in resource_name is a snake case representation of multiple words. If it is one word, it replaces with nothing.

Examples

iex > template = “<%= path_for_type %>” iex > Braise.AdapterTemplate.replace_path_for_type_variable(template, “patient”) “” iex > Braise.AdapterTemplate.replace_path_for_type_variable(template, “staff_members”) “pathForType: function(type) { var decamelized = Ember.String.decamelize(type); return Ember.String.pluralize(decamelized); }, “

replace_template_variables(template_string, resource)

replace_uri_variables(template, arg2)

Replaces the scheme, host and path template variables with the representations from the Braise.Resource.url

Examples

iex > template = “<%= scheme %> <%= host %> <%= path %>” iex > Braise.AdapterTemplate.replace_uri_variables(template, URI.parse(“http://bizdev.com/api”)) “http bizdev.com /api”