braise v0.3.4 Braise.Model

Representation of a RESTful resource with a name plus a collection of attributes.

Functions

parse_from_resource(resource)

Returns a Braise Model with a collection of attributes assigned to the resource name.

Examples

iex > resource = %Braise.Resource{name: "pirate", definition: %{"name" => "booty"}, response: %{"name" => "name"}}
iex > Braise.Model.parse_from_resource(resource)
%Braise.Model{name: "pirate", attributes: [%Braise.Attribute{name: "name", type: null, format: null}]}