braise v0.3.3 Braise.AdapterTemplate

Constructs a string representation of an Ember-CLI Adapter given the passed in JSON Schema Resource.

Functions

generate_from_resource(resource)

path_for_type(resource_name)

Returns javascript string for pathForType property if it is the passed in resource_name is a snake case representation of multiple words. If it is one word, it replaces with nothing.

Examples

iex > template = “<%= path_for_type %>” iex > Braise.AdapterTemplate.path_for_type(“patient”) “” iex > Braise.AdapterTemplate.path_for_type(“staff_members”) “pathForType: function(type) { var underscorized = Ember.String.underscore(type); return Ember.String.pluralize(underscorized); }, “